backend/requirements.txt
Flask==2.0.1
flask-cors==3.0.10
flask-mysqldb==0.2.0
PyJWT==2.1.0
Werkzeug==2.0.1 